added game logic file and moved socket logic from server.js
